requirements.txt
# fastapi<=0.111.0,>=0.111.0
# uvicorn<=0.30.1,>=0.30.1
# sqlalchemy>=2.0.29,<=2.0.30
# pymysql>=1.1.1,<=1.1.1
# SQLAlchemy>=2.0.29,<=2.0.29
# Mako>=1.3.5,<=1.3.5
# alembic>=1.13.2,<=1.13.2
# email_validator>=2.2.0,<2.2.0
# bcrypt>=4.1.3,<4.1.3
# PyJWT>=2.8.0,<2.8.0

fastapi<=0.111.0,>=0.111.0
uvicorn<=0.30.1,>=0.30.1
sqlalchemy>=2.0.29,<=2.0.30
pymysql>=1.1.1,<=1.1.1
SQLAlchemy>=2.0.29,<=2.0.29
Mako>=1.3.5,<=1.3.5
alembic>=1.13.2,<=1.13.2
email_validator>=2.2.0,<=2.2.0
bcrypt>=4.1.3,<=4.1.3
PyJWT>=2.8.0,<=2.8.0
python-jose>=3.3.0
passlib==1.7.4
cryptography==41.0.7
